feat: add an array of refs
If several external clicks are used in one component,
then when you click on the drop-down
element,
it will be closed due to the fact
that one of them contains the target element,
and the
other does not, and the closing condition is triggered.
To avoid this, pass an array of links to the
trap
and check internally that all links do not contain a target,
if at least one link contains a
target, then the closure will not occur.