feat(components): use prediction data
Create an action query to get aggregated price data for a week.
Put the resulting data into the
redux store.
Use this data to display pricing information for the coming days.
The functions
responsible for converting dates in the ticket
component have been moved to utilities because they

are to be used in the forecasting component.

МЕТА ДАННЫЕ: close #128